Refactor security and encryption package structure
- Move `PasswordUtils` and migration logic to `dev.leonlatsch.photok.encryption` package.
- Rename `UnlockCipherUseCase` to `UnlockBiometricCipherPrompt` and move to `encryption.ui`.
- Remove `EncryptionModule` and `BiometricUnlock`.
- Update package references in manifests, navigation graphs, and UI components.
- Comment out debug suffix logic in `build.gradle.kts`.
