+++ app/src/main/java/dev/leonlatsch/photok/encryption/migration/MigrationService.kt
@@ -1,5 +1,5 @@
 /*
- *   Copyright 2020–2026 Leon Latsch
+ *   Copyright 2020-2026 Leon Latsch
  *
  *   Licensed under the Apache License, Version 2.0 (the "License");
  *   you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  *   limitations under the License.
  */
 
-package dev.leonlatsch.photok.security.migration
+package dev.leonlatsch.photok.encryption.migration
 
 import android.annotation.SuppressLint
 import android.app.Notification
