+++ app/src/main/java/dev/leonlatsch/photok/encryption/migration/ui/EncryptionMigrationScreenMigrating.kt
@@ -14,23 +14,28 @@
  *   limitations under the License.
  */
 
-package dev.leonlatsch.photok.security.migration.ui
+package dev.leonlatsch.photok.encryption.migration.ui
 
+import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
+import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.PreviewLightDark
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
@@ -38,9 +43,11 @@ import dev.leonlatsch.photok.R
 import dev.leonlatsch.photok.ui.components.AppName
 import dev.leonlatsch.photok.ui.theme.AppTheme
 
+val IconSize = 72.dp
+
 @Composable
-fun EncryptionMigrationScreenSuccess(
-    uiState: LegacyEncryptionMigrationUiState.Success
+fun EncryptionMigrationScreenMigrating(
+    uiState: LegacyEncryptionMigrationUiState.Migrating,
 ) {
     Scaffold { contentPadding ->
         Box(
@@ -60,7 +67,7 @@ fun EncryptionMigrationScreenSuccess(
                 AppName()
 
                 Text(
-                    text = stringResource(R.string.migration_done_title),
+                    text = stringResource(R.string.migration_running_title),
                     fontSize = 22.sp,
                     fontWeight = FontWeight.Bold,
                 )
@@ -70,25 +77,62 @@ fun EncryptionMigrationScreenSuccess(
                     verticalArrangement = Arrangement.spacedBy(12.dp),
                 ) {
                     LoadingIndicator(
-                        progressPercentage = 1f
+                        progressPercentage = uiState.progressPercentage
                     )
 
                     Text(
-                        text = stringResource(R.string.migration_done_progress),
+                        text = stringResource(
+                            R.string.migration_running_progress,
+                            uiState.processedFiles,
+                            uiState.totalFiles
+                        ),
                         color = MaterialTheme.colorScheme.secondary,
                     )
                 }
             }
+
+            Text(
+                modifier = Modifier
+                    .align(Alignment.BottomCenter)
+                    .padding(16.dp),
+                text = stringResource(R.string.migration_running_advice),
+                textAlign = TextAlign.Center,
+                color = MaterialTheme.colorScheme.outline,
+            )
         }
+    }
+}
+
+@Composable
+fun LoadingIndicator(progressPercentage: Float, modifier: Modifier = Modifier) {
+    Box(modifier = modifier) {
+        Image(
+            modifier = Modifier
+                .align(Alignment.Center)
+                .size(IconSize),
+            painter = painterResource(R.drawable.app_icon),
+            contentDescription = null
+        )
+
+        CircularProgressIndicator(
+            progress = { progressPercentage },
+            modifier = Modifier
+                .align(Alignment.Center)
+                .size(IconSize + 40.dp),
+            strokeWidth = 10.dp,
+        )
     }
 }
 
 @PreviewLightDark
 @Composable
-private fun PreviewSuccess() {
+private fun Preview() {
     AppTheme {
-        EncryptionMigrationScreenSuccess(
-            uiState = LegacyEncryptionMigrationUiState.Success
+        EncryptionMigrationScreenMigrating(
+            uiState = LegacyEncryptionMigrationUiState.Migrating(
+                totalFiles = 100,
+                processedFiles = 40,
+            ),
         )
     }
 }
