+++ app/src/main/java/dev/leonlatsch/photok/settings/ui/changepassword/ChangePasswordDialog.kt
@@ -24,10 +24,10 @@ import dagger.hilt.android.AndroidEntryPoint
 import dev.leonlatsch.photok.BR
 import dev.leonlatsch.photok.R
 import dev.leonlatsch.photok.databinding.DialogChangePasswordBinding
+import dev.leonlatsch.photok.encryption.domain.PasswordUtils
 import dev.leonlatsch.photok.other.extensions.empty
 import dev.leonlatsch.photok.other.extensions.hide
 import dev.leonlatsch.photok.other.extensions.show
-import dev.leonlatsch.photok.security.PasswordUtils
 import dev.leonlatsch.photok.uicomponnets.bindings.BindableDialogFragment
 
 /**
